Changed yml file, added resnet model for slimming , doubt in resnet batchnorm
