+++ trailmet/models/__init__.py
@@ -1,8 +1,9 @@
 from .resnet import get_resnet_model
+from .resnet import *
 
 from .vgg_16_ns import *
-from .Resnet_164_ns import *
-class ModelsFactory(object):
+class ModelsFactory(obje
+                    ct):
     @staticmethod
     def create_model(name, num_classes=100, pretrained = False, **kwargs):
         """
